requirements.txt
decord
diffusers
imageio
einops
gdown
numpy
opencv_python
Pillow
safetensors
torch==2.4.0
torchvision
torchaudio
tqdm
lpips
wandb
accelerate
pandas>=1.3.0
matplotlib>=3.3.0
scipy>=1.7.0